backend/requirements.txt
# Core Django Framework
Django==5.0.4

# Django Extensions
djangorestframework==3.15.1
django-filter==24.2
django-nextjs==2.1.3
django-cors-headers==4.3.1

# Database Driver
psycopg2-binary==2.9.9

# Async/WebSocket Support (if needed)
channels==4.1.0

# HTTP Client (only if you make external API calls)
requests==2.31.0
